requirements.txt
requests==2.28.1
urllib3==1.26.10
google-cloud-storage==2.4.0
moviepy==1.0.3